requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
asyncer==0.0.7
asyncpg==0.30.0
attrs==25.3.0
backoff==2.2.1
backports.tarfile==1.2.0
beautifulsoup4==4.13.4
bidict==0.23.1
build==1.2.2.post1
CacheControl==0.14.3
certifi==2025.4.26
chainlit==2.5.5
charset-normalizer==3.4.2
chevron==0.14.0
cleo==2.1.0
click==8.1.8
colorama==0.4.6
crashtest==0.4.1
dataclasses-json==0.6.7
deepl==1.22.0
Deprecated==1.2.18
distlib==0.3.9
distro==1.9.0
dotenv==0.9.9
dulwich==0.22.8
fastapi==0.115.12
fastjsonschema==2.21.1
feedparser==6.0.11
filelock==3.18.0
filetype==1.2.0
findpython==0.6.3
frozenlist==1.6.0
fsspec==2025.3.2
google==3.0.0
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.0
inflection==0.5.1
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
keyring==25.6.0
kiwipiepy==0.21.0
kiwipiepy_model==0.21.0
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.58
langchain-openai==0.3.14
langchain-postgres==0.0.14
langchain-teddynote==0.3.45
langchain-text-splitters==0.3.8
langgraph==0.4.3
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langsmith==0.3.37
Lazify==0.4.0
literalai==0.1.201
load-dotenv==0.1.0
lz4==4.4.4
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.6.0
mmh3==4.1.0
monotonic==1.6
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
olefile==0.47
openai==1.76.0
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-alephalpha==0.39.3
opentelemetry-instrumentation-anthropic==0.39.3
opentelemetry-instrumentation-bedrock==0.39.3
opentelemetry-instrumentation-chromadb==0.39.3
opentelemetry-instrumentation-cohere==0.39.3
opentelemetry-instrumentation-crewai==0.39.3
opentelemetry-instrumentation-google-generativeai==0.39.3
opentelemetry-instrumentation-groq==0.39.3
opentelemetry-instrumentation-haystack==0.39.3
opentelemetry-instrumentation-lancedb==0.39.3
opentelemetry-instrumentation-langchain==0.39.3
opentelemetry-instrumentation-llamaindex==0.39.3
opentelemetry-instrumentation-logging==0.52b1
opentelemetry-instrumentation-marqo==0.39.3
opentelemetry-instrumentation-milvus==0.39.3
opentelemetry-instrumentation-mistralai==0.39.3
opentelemetry-instrumentation-ollama==0.39.3
opentelemetry-instrumentation-openai==0.39.3
opentelemetry-instrumentation-pinecone==0.39.3
opentelemetry-instrumentation-qdrant==0.39.3
opentelemetry-instrumentation-replicate==0.39.3
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-sagemaker==0.39.3
opentelemetry-instrumentation-sqlalchemy==0.52b1
opentelemetry-instrumentation-threading==0.52b1
opentelemetry-instrumentation-together==0.39.3
opentelemetry-instrumentation-transformers==0.39.3
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-vertexai==0.39.3
opentelemetry-instrumentation-watsonx==0.39.3
opentelemetry-instrumentation-weaviate==0.39.3
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-semantic-conventions-ai==0.4.3
opentelemetry-util-http==0.52b1
orjson==3.10.16
ormsgpack==1.9.1
packaging>=24.0,<25.0
pandas==2.2.3
pbs-installer==2025.4.9
pdf2image==1.17.0
pexpect==4.9.0
pgvector==0.3.6
pillow==11.2.1
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
pinecone-text==0.10.0
pkginfo==1.12.1.2
platformdirs==4.3.7
poetry==2.1.3
poetry-core==2.1.3
poetry-plugin-shell==1.0.1
posthog==3.25.0
propcache==0.3.1
protobuf==5.29.4
protoc-gen-openapiv2==0.0.1
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg-pool==3.2.6
ptyprocess==0.7.0
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
PyJWT==2.10.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-multipart==0.0.18
python-socketio==5.13.0
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
rank-bm25==0.2.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==80.3.1
sgmllib3k==1.0.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
sqlmodel==0.0.24
sse-starlette==2.3.3
starlette==0.41.3
sympy==1.14.0
syncer==2.0.3
tavily-python==0.7.2
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.13.2
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
traceloop-sdk==0.39.3
trove-classifiers==2025.5.1.12
types-requests==2.32.0.20250328
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uptrace==1.31.0
urllib3==2.4.0
uvicorn==0.34.2
virtualenv==20.31.1
watchfiles==0.20.0
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
